Feature update auth
we can use tier in user part

+++ back/lacheln/user/src/main/java/aba3/lucid/cart/business/CartBusiness.java
@@ -108,13 +108,19 @@ public class CartBusiness {
     }
     
     // 장바구니 상품제거
-    public API<String> deleteCart(CartDeleteRequest cartDeleteRequest)    {
-        cartService.removeCart(cartDeleteRequest.getCartIds());
+    public API<String> deleteCart(CartDeleteRequest request)    {
+        if(request == null) {
+            throw new ApiException(ErrorCode.BAD_REQUEST, "장바구니가 없습니다.");
+        }
+        cartService.removeCart(request.getCartIds());
         return API.OK("장바구니 상품 제거 성공..");
     }
 
     // 장바구니 비우기
     public API<String> deleteAllCart(String userId)  {
+        if(userId == null) {
+            throw new ApiException(ErrorCode.NULL_POINT, "해당하는 유저가 없거나, 장바구니가 비어있습니다.");
+        }
         cartService.removeAllCart(userId);
         return API.OK("장바구니 비우기 성공..");
     }
